refactor: ts fixes
