+++ src/lib/characterPreview/ShowcaseCustomizationSidebar.tsx
@@ -4,8 +4,8 @@ import { AggregationColor } from 'antd/es/color-picker/color'
 import { GlobalToken } from 'antd/lib/theme/interface'
 import { usePublish } from 'hooks/usePublish'
 import { ShowcaseSource } from 'lib/characterPreview/CharacterPreviewComponents'
-import { useAsyncSimScoringExecution } from 'lib/characterPreview/CharacterStatSummary'
 import { DEFAULT_SHOWCASE_COLOR, editShowcasePreferences } from 'lib/characterPreview/showcaseCustomizationController'
+import { useAsyncSimScoringExecution } from 'lib/characterPreview/UseAsyncSimScoringExecution'
 import { ShowcaseColorMode, Stats } from 'lib/constants/constants'
 import { SavedSessionKeys } from 'lib/constants/constantsSession'
 import { Assets } from 'lib/rendering/assets'
