+++ src/lib/characterPreview/ShowcaseDpsScore.tsx
@@ -3,8 +3,8 @@ import { Button, Card, ConfigProvider, Flex, Segmented } from 'antd'
 import type { GlobalToken } from 'antd/es/theme/interface'
 import { CharacterCardCombatStats } from 'lib/characterPreview/CharacterCardCombatStats'
 import { OverlayText, showcaseOutline } from 'lib/characterPreview/CharacterPreviewComponents'
-import { useAsyncSimScoringExecution } from 'lib/characterPreview/CharacterStatSummary'
 import StatText from 'lib/characterPreview/StatText'
+import { useAsyncSimScoringExecution } from 'lib/characterPreview/UseAsyncSimScoringExecution'
 import { CUSTOM_TEAM, DEFAULT_TEAM, SETTINGS_TEAM } from 'lib/constants/constants'
 import { defaultGap } from 'lib/constants/constantsUi'
 import { SingleRelicByPart } from 'lib/gpu/webgpuTypes'
